# backend/contracts/openapi/v1/provider.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "TAKEIN Provider API",
    "version": "1.0.0",
    "description": "Existing provider-owner and provider-branch API contract. Every operation requires JWT authentication plus the server-side provider account, branch, and menu permission checks identified by x-provider-permission."
  },
  "servers": [
    {
      "url": "https://api.takein.id/api",
      "description": "Production API ingress"
    }
  ],
  "paths": {
...
    "/provider/profile/password": {
      "put": {
        "operationId": "updateProviderPassword",
        "summary": "Change the current provider account password",
        "x-authentication": "bearerAuth",
        "x-provider-permission": "profile",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderPasswordUpdateRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "JsonObject": {
        "type": "object",
        "additionalProperties": true
      },
...
      "ProviderPasswordUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "current_password",
          "password",
          "password_confirmation"
        ],
        "properties": {
          "current_password": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "password_confirmation": {
            "type": "string",
            "minLength": 8
          }
        }
      },
...
    }
  }
}